refactor: divide main in to two
